api/requirements.txt
# filepath: requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-exporter-otlp==1.20.0
